+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/abstracts/GroupEventMessage.kt
@@ -1,3 +1,7 @@
 package dev.inmo.tgbotapi.types.message.abstracts
 
-interface GroupEventMessage : ChatEventMessage, FromUserMessage
+import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.GroupEvent
+
+interface GroupEventMessage : ChatEventMessage, FromUserMessage {
+    override val chatEvent: GroupEvent
+}